Move fundingTrades sync-schema to separate file
